Updated New Test Scenarios
